requirements.txt
# requirements.txt

# GUI Framework (Tkinter)
customtkinter

# Application permissions
pyuac
requests

# Cryptography (for file encryption/decryption)
cryptography

# Antivirus-related libraries
pyclamd
pywin32

#Typography
Pillow
googlefonts-installer
